requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosmtplib==4.0.2
annotated-types==0.7.0
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
flake8==7.3.0
frozenlist==1.7.0
idna==3.10
mccabe==0.7.0
multidict==6.6.4
mypy==1.18.2
mypy_extensions==1.1.0
pathspec==0.12.1
propcache==0.3.2
pycodestyle==2.14.0
pydantic==2.11.9
pydantic_core==2.33.2
pyflakes==3.4.0
python-dotenv==1.1.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
yarl==1.20.1
